+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/offline/OfflineDataset.kt
@@ -1,4 +1,4 @@
-package com.github.HumanLearning2021.HumanLearningApp.room
+package com.github.HumanLearning2021.HumanLearningApp.offline
 
 import com.github.HumanLearning2021.HumanLearningApp.model.Category
 import com.github.HumanLearning2021.HumanLearningApp.model.Dataset
