+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/room/RoomDatabaseDefinition.kt
@@ -1,11 +1,10 @@
 package com.github.HumanLearning2021.HumanLearningApp.room
 
-import androidx.room.*
-import com.github.HumanLearning2021.HumanLearningApp.model.Category
-import com.github.HumanLearning2021.HumanLearningApp.model.User
+import androidx.room.Database
+import androidx.room.RoomDatabase
 
 @Database(entities = [RoomCategory::class, RoomCategorizedPicture::class, RoomDataset::class, RoomUser::class], version = 1)
-abstract class OfflineDatabase : RoomDatabase() {
+abstract class RoomDatabaseDefinition : RoomDatabase() {
     abstract fun categoryDao(): CategoryDao
     abstract fun categorizedPictureDao(): CategorizedPictureDao
     abstract fun datasetDao(): DatasetDao
